+++ src/sphinx_codeautolink/__init__.py
@@ -14,7 +14,7 @@ def setup(app: Sphinx):
     state = SphinxCodeAutoLink()
     app.setup_extension('sphinx.ext.autodoc')
     app.add_css_file('sphinx-codeautolink.css')
-    app.add_config_value('codeautolink_autodoc_inject', True, 'html', types=[bool])
+    app.add_config_value('codeautolink_autodoc_inject', False, 'html', types=[bool])
     app.add_config_value('codeautolink_global_preface', '', 'html', types=[str])
     app.add_config_value('codeautolink_custom_blocks', {}, 'html', types=[dict])
     app.add_config_value('codeautolink_concat_default', False, 'html', types=[bool])